Split the YAML stuff out into its own module
