+++ syntax/src/yaml/mod.rs
@@ -1,0 +1,5 @@
+mod analysis;
+mod types;
+
+pub use analysis::analyse;
+pub use types::*;
